+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## v2.0.1
+## v2.1.0
 
+- Added `maximumBreadth` option to limit stringification at a specific object or array "width" (number of properties / values)
+- Added `maximumDepth` option to limit stringification at a specific nesting depth
 - Fixed maximum spacer length (10)
 - Fixed TypeScript definition